Share more mock types in std and WTF across tests.
