+++ clang/test/Analysis/Checkers/WebKit/unchecked-lambda-captures.cpp
@@ -2,186 +2,6 @@
 
 #include "mock-types.h"
 
-namespace std {
-
-template <typename T>
-T&& move(T& t) {
-  return static_cast<T&&>(t);
-}
-
-namespace ranges {
-
-template<typename IteratorType, typename CallbackType>
-void for_each(IteratorType first, IteratorType last, CallbackType callback) {
-  for (auto it = first; !(it == last); ++it)
-    callback(*it);
-}
-
-struct all_of_impl {
-  template <typename Collection, typename Predicate>
-  constexpr bool operator()(const Collection& collection, Predicate predicate) const {
-    for (auto it = collection.begin(); it != collection.end(); ++it) {
-      if (!predicate(*it))
-        return false;
-    }
-    return true;
-  }
-};
-inline constexpr auto all_of = all_of_impl {};
-
-}
-
-}
-
-namespace WTF {
-
-namespace Detail {
-
-template<typename Out, typename... In>
-class CallableWrapperBase {
-public:
-    virtual ~CallableWrapperBase() { }
-    virtual Out call(In...) = 0;
-};
-
-template<typename, typename, typename...> class CallableWrapper;
-
-template<typename CallableType, typename Out, typename... In>
-class CallableWrapper : public CallableWrapperBase<Out, In...> {
-public:
-    explicit CallableWrapper(CallableType& callable)
-        : m_callable(callable) { }
-    Out call(In... in) final { return m_callable(in...); }
-
-private:
-    CallableType m_callable;
-};
-
-} // namespace Detail
-
-template<typename> class Function;
-
-template<typename Out, typename... In> Function<Out(In...)> adopt(Detail::CallableWrapperBase<Out, In...>*);
-
-template <typename Out, typename... In>
-class Function<Out(In...)> {
-public:
-    using Impl = Detail::CallableWrapperBase<Out, In...>;
-
-    Function() = default;
-
-    template<typename FunctionType>
-    Function(FunctionType f)
-        : m_callableWrapper(new Detail::CallableWrapper<FunctionType, Out, In...>(f)) { }
-
-    Out operator()(In... in) const { return m_callableWrapper->call(in...); }
-    explicit operator bool() const { return !!m_callableWrapper; }
-
-private:
-    enum AdoptTag { Adopt };
-    Function(Impl* impl, AdoptTag)
-        : m_callableWrapper(impl)
-    {
-    }
-
-    friend Function adopt<Out, In...>(Impl*);
-
-    std::unique_ptr<Impl> m_callableWrapper;
-};
-
-template<typename Out, typename... In> Function<Out(In...)> adopt(Detail::CallableWrapperBase<Out, In...>* impl)
-{
-    return Function<Out(In...)>(impl, Function<Out(In...)>::Adopt);
-}
-
-template <typename KeyType, typename ValueType>
-class HashMap {
-public:
-  HashMap();
-  HashMap([[clang::noescape]] const Function<ValueType()>&);
-  void ensure(const KeyType&, [[clang::noescape]] const Function<ValueType()>&);
-  bool operator+([[clang::noescape]] const Function<ValueType()>&) const;
-  static void ifAny(HashMap, [[clang::noescape]] const Function<bool(ValueType)>&);
-
-private:
-  ValueType* m_table { nullptr };
-};
-
-class ScopeExit final {
-public:
-  template<typename ExitFunctionParameter>
-  explicit ScopeExit(ExitFunctionParameter&& exitFunction)
-    : m_exitFunction(std::move(exitFunction)) {
-  }
-
-  ScopeExit(ScopeExit&& other)
-    : m_exitFunction(std::move(other.m_exitFunction))
-    , m_execute(std::move(other.m_execute)) {
-  }
-
-  ~ScopeExit() {
-    if (m_execute)
-      m_exitFunction();
-  }
-
-  WTF::Function<void()> take() {
-    m_execute = false;
-    return std::move(m_exitFunction);
-  }
-
-  void release() { m_execute = false; }
-
-  ScopeExit(const ScopeExit&) = delete;
-  ScopeExit& operator=(const ScopeExit&) = delete;
-  ScopeExit& operator=(ScopeExit&&) = delete;
-
-private:
-  WTF::Function<void()> m_exitFunction;
-  bool m_execute { true };
-};
-
-template<typename ExitFunction> ScopeExit makeScopeExit(ExitFunction&&);
-template<typename ExitFunction>
-ScopeExit makeScopeExit(ExitFunction&& exitFunction)
-{
-    return ScopeExit(std::move(exitFunction));
-}
-
-// Visitor adapted from http://stackoverflow.com/questions/25338795/is-there-a-name-for-this-tuple-creation-idiom
-
-template<class A, class... B> struct Visitor : Visitor<A>, Visitor<B...> {
-    Visitor(A a, B... b)
-        : Visitor<A>(a)
-        , Visitor<B...>(b...)
-    {
-    }
-
-    using Visitor<A>::operator ();
-    using Visitor<B...>::operator ();
-};
-
-template<class A> struct Visitor<A> : A {
-    Visitor(A a)
-        : A(a)
-    {
-    }
-
-    using A::operator();
-};
-
-template<class... F> Visitor<F...> makeVisitor(F... f)
-{
-    return Visitor<F...>(f...);
-}
-
-void opaqueFunction();
-template <typename Visitor, typename... Variants> void visit(Visitor&& v, Variants&&... values)
-{
-  opaqueFunction();
-}
-
-} // namespace WTF
-
 struct A {
   static void b();
 };
