+++ clang/test/Analysis/Checkers/WebKit/unretained-lambda-captures.mm
@@ -1,110 +1,7 @@
 // RUN: %clang_analyze_cc1 -analyzer-checker=alpha.webkit.UnretainedLambdaCapturesChecker -verify %s
 
+#include "mock-types.h"
 #include "objc-mock-types.h"
-
-namespace std {
-
-template <typename T>
-class unique_ptr {
-private:
-  T *t;
-
-public:
-  unique_ptr() : t(nullptr) { }
-  unique_ptr(T *t) : t(t) { }
-  ~unique_ptr() {
-    if (t)
-      delete t;
-  }
-  template <typename U> unique_ptr(unique_ptr<U>&& u)
-    : t(u.t)
-  {
-    u.t = nullptr;
-  }
-  T *get() const { return t; }
-  T *operator->() const { return t; }
-  T &operator*() const { return *t; }
-  unique_ptr &operator=(T *) { return *this; }
-  explicit operator bool() const { return !!t; }
-};
-
-};
-
-namespace WTF {
-
-namespace Detail {
-
-template<typename Out, typename... In>
-class CallableWrapperBase {
-public:
-    virtual ~CallableWrapperBase() { }
-    virtual Out call(In...) = 0;
-};
-
-template<typename, typename, typename...> class CallableWrapper;
-
-template<typename CallableType, typename Out, typename... In>
-class CallableWrapper : public CallableWrapperBase<Out, In...> {
-public:
-    explicit CallableWrapper(CallableType& callable)
-        : m_callable(callable) { }
-    Out call(In... in) final { return m_callable(in...); }
-
-private:
-    CallableType m_callable;
-};
-
-} // namespace Detail
-
-template<typename> class Function;
-
-template<typename Out, typename... In> Function<Out(In...)> adopt(Detail::CallableWrapperBase<Out, In...>*);
-
-template <typename Out, typename... In>
-class Function<Out(In...)> {
-public:
-    using Impl = Detail::CallableWrapperBase<Out, In...>;
-
-    Function() = default;
-
-    template<typename FunctionType>
-    Function(FunctionType f)
-        : m_callableWrapper(new Detail::CallableWrapper<FunctionType, Out, In...>(f)) { }
-
-    Out operator()(In... in) const { return m_callableWrapper->call(in...); }
-    explicit operator bool() const { return !!m_callableWrapper; }
-
-private:
-    enum AdoptTag { Adopt };
-    Function(Impl* impl, AdoptTag)
-        : m_callableWrapper(impl)
-    {
-    }
-
-    friend Function adopt<Out, In...>(Impl*);
-
-    std::unique_ptr<Impl> m_callableWrapper;
-};
-
-template<typename Out, typename... In> Function<Out(In...)> adopt(Detail::CallableWrapperBase<Out, In...>* impl)
-{
-    return Function<Out(In...)>(impl, Function<Out(In...)>::Adopt);
-}
-
-template <typename KeyType, typename ValueType>
-class HashMap {
-public:
-  HashMap();
-  HashMap([[clang::noescape]] const Function<ValueType()>&);
-  void ensure(const KeyType&, [[clang::noescape]] const Function<ValueType()>&);
-  bool operator+([[clang::noescape]] const Function<ValueType()>&) const;
-  static void ifAny(HashMap, [[clang::noescape]] const Function<bool(ValueType)>&);
-
-private:
-  ValueType* m_table { nullptr };
-};
-
-} // namespace WTF
 
 struct A {
   static void b();
